add utils package as dependency best practice
