+++ v2/route.go
@@ -3,6 +3,8 @@ package v2
 import (
 	"strings"
 
+	"github.com/pixeltopic/sayori/v2/utils"
+
 	"context"
 )
 
@@ -181,19 +183,19 @@ func createHandlerFunc(route *Route) handlerFunc {
 	return func(ctx context.Context) {
 		var (
 			ok  bool
-			msg = GetMsg(ctx)
+			msg = utils.GetMsg(ctx)
 			cmd = msg.Content
 		)
 
 		prefix := route.getGuildPrefix(msg.GuildID)
-		ctx = WithPrefix(ctx, prefix)
+		ctx = utils.WithPrefix(ctx, prefix)
 		if cmd, ok = trimPrefix(cmd, prefix); !ok {
 			return
 		}
 
 		args, err := handleParse(route.c, cmd)
 		if err != nil {
-			ctx = WithErr(ctx, err)
+			ctx = utils.WithErr(ctx, err)
 			route.c.Resolve(ctx)
 			return
 		}
@@ -203,18 +205,18 @@ func createHandlerFunc(route *Route) handlerFunc {
 			return
 		}
 
-		ctx = WithAlias(ctx, args[:depth])
-		ctx = WithArgs(ctx, args[depth:])
+		ctx = utils.WithAlias(ctx, args[:depth])
+		ctx = utils.WithArgs(ctx, args[depth:])
 
 		if err = handleMiddlewares(ctx, route.middlewares); err != nil {
-			ctx = WithErr(ctx, err)
+			ctx = utils.WithErr(ctx, err)
 			route.c.Resolve(ctx)
 			return
 		}
 
 		err = route.c.Handle(ctx)
 		if err != nil {
-			ctx = WithErr(ctx, err)
+			ctx = utils.WithErr(ctx, err)
 		}
 
 		route.c.Resolve(ctx)
